+++ core/libs/GPU/cmtkImageSymmetryPlaneFunctionalDevice.cxx
@@ -2,7 +2,7 @@
 //
 //  Copyright 1997-2009 Torsten Rohlfing
 //
-//  Copyright 2004-2010 SRI International
+//  Copyright 2004-2012 SRI International
 //
 //  This file is part of the Computational Morphometry Toolkit.
 //
@@ -86,7 +86,7 @@ ImageSymmetryPlaneFunctionalDevice::Evaluate()
     {
     for ( size_t i = 0; i < 3; ++i )
       {
-      matrix[j][i] /= this->m_Volume->Size[i];
+      matrix[j][i] /= static_cast<float>( this->m_Volume->Size[i] );
       }
     }
